split grid from bulk page
